+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -571,7 +571,7 @@ class HTTPTool(BaseModel):
     http: HTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6A6CKCVAAMX0ZZDQH3NN3NK"
+        "01K6AKQCPXKBE0QCRMSN5AN41E"
     )
 
     requires_approval: Optional[bool] = False