+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6A6CKCZ5RZA1T4FPSK916V2"
+        "01K6AKQCQ1BR7B5VTQATY7R7ZS"
     )
 
     requires_approval: Optional[bool] = False