+++ packages/orq-rc/RELEASES.md
@@ -2959,3 +2959,13 @@ Based on:
 - [python v3.13.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.43] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.43 - packages/orq-rc
+
+## 2025-09-29 11:30:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.4 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.44] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.44] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.44 - packages/orq-rc